requirements.txt
gymnasium
numpy
neat-python
networkx
matplotlib